+++ src/CUITe.IntegrationTests/ObjectRecorder/Features/Recording.feature.custom.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace CUITe.IntegrationTests.ObjectRecorder.Features
+{
+#if DEBUG
+    [DeploymentItem(@".\CUITe_ObjectRecorder\bin\Debug\CUITe_ObjectRecorder.exe")]
+#else
+    [DeploymentItem(@".\CUITe_ObjectRecorder\bin\Release\CUITe_ObjectRecorder.exe")]
+#endif
+    public partial class RecordingFeature
+    {
+    }
+}